+++ src/constant.ts
@@ -1,0 +1,53 @@
+/**
+ * One unit of milliseconds.
+ */
+
+export const MILLISECOND = 1;
+
+/**
+ * One second in milliseconds.
+ */
+
+export const SECOND = 1_000;
+
+/**
+ * One minute in milliseconds.
+ */
+
+export const MINUTE = 60_000;
+
+/**
+ * One hour in milliseconds.
+ */
+
+export const HOUR = 3_600_000;
+
+/**
+ * One day in milliseconds.
+ */
+
+export const DAY = 86_400_000;
+
+/**
+ * Number of milliseconds in a second.
+ */
+
+export const MILLISECONDS_IN_A_SECOND = SECOND / MILLISECOND;
+
+/**
+ * Number of seconds in a minute.
+ */
+
+export const SECONDS_IN_MINUTE = MINUTE / SECOND;
+
+/**
+ * Number of minutes in an hour.
+ */
+
+export const MINUTES_IN_HOUR = HOUR / MINUTE;
+
+/**
+ * Number of hours in a day.
+ */
+
+export const HOURS_IN_DAY = DAY / HOUR;